API'S + Métiers

+++ src/GUI/PostBackController.java
@@ -6,6 +6,7 @@
 package GUI;
 
 import Entities.Post;
+import Entities.Rating;
 import Services.ServicePost;
 import java.awt.event.ActionEvent;
 import java.io.File;
@@ -15,7 +16,11 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
+import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.value.ObservableValue;
+import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -25,9 +30,13 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonBar;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.Dialog;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableColumn.CellDataFeatures;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
@@ -36,6 +45,7 @@ import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
+import javafx.util.Callback;
 
 /**
  * FXML Controller class
@@ -47,8 +57,7 @@ public class PostBackController implements Initializable {
    
     @FXML
     private TableView<Post> tftableview;
-    @FXML
-    private TableColumn<Post, Integer> tcpost;
+   
     @FXML
     private TableColumn<Post, String> tctheme;
     @FXML
@@ -73,6 +82,15 @@ public class PostBackController implements Initializable {
     private TableColumn<Post, String> images;
     @FXML
     private Button tfadd;
+    @FXML
+    private Button tfadd1;
+    @FXML
+    private TableColumn<Post, Integer> tclike;
+    @FXML
+    private TableColumn<Post, Integer> tcdislike;
+    @FXML
+    private TextField searchField;
+    
 
     
     /**
@@ -83,7 +101,7 @@ public class PostBackController implements Initializable {
           ServicePost service = new ServicePost();
          ObservableList<Post> list = service.displayPosts();
           //System.out.print(list);
-          tcpost.setCellValueFactory(new PropertyValueFactory<Post, Integer>("id"));
+         
           tctheme.setCellValueFactory(new PropertyValueFactory<Post, String>("Theme"));
           tcnom.setCellValueFactory(new PropertyValueFactory<Post, String>("nom"));
           tcdate.setCellValueFactory(new PropertyValueFactory<Post, Date>("date_Creation"));
@@ -116,7 +134,31 @@ public class PostBackController implements Initializable {
                     }
                 };
             });
+           // tccmnts.setCellValueFactory(new PropertyValueFactory<Post, Integer>("commentCount"));
+          tclike.setCellValueFactory(new Callback<CellDataFeatures<Post, Integer>, ObservableValue<Integer>>() {
+            @Override
+            public ObservableValue<Integer> call(CellDataFeatures<Post, Integer> param) {
+                int postId = param.getValue().getId();
+                int likeCount = service.getLikeCount(postId);
+                return new SimpleIntegerProperty(likeCount).asObject();
+            }
+        });
 
+        tcdislike.setCellValueFactory(new Callback<CellDataFeatures<Post, Integer>, ObservableValue<Integer>>() {
+           @Override
+           public ObservableValue<Integer> call(CellDataFeatures<Post, Integer> param) {
+               int postId = param.getValue().getId();
+               int dislikeCount = service.getDislikeCount(postId);
+               return new SimpleIntegerProperty(dislikeCount).asObject();
+           }
+       });   
+           
+        // add an event listener to the search field
+    searchField.textProperty().addListener((observable, oldValue, newValue) -> {
+        searchPosts();
+    });
+    
+           
           tftableview.setItems(list);
     }    
        
@@ -124,7 +166,6 @@ public class PostBackController implements Initializable {
     {
         ServicePost service=new ServicePost();
          ObservableList<Post> list = service.displayPosts();
-         tcpost.setCellValueFactory(new PropertyValueFactory<Post, Integer>("id"));
           tctheme.setCellValueFactory(new PropertyValueFactory<Post, String>("Theme"));
           tcnom.setCellValueFactory(new PropertyValueFactory<Post, String>("nom"));
           tcdate.setCellValueFactory(new PropertyValueFactory<Post, Date>("date_Creation"));
@@ -154,7 +195,32 @@ public class PostBackController implements Initializable {
                     }
                 };
             });
-          tftableview.setItems(list);
+           //  tccmnts.setCellValueFactory(new PropertyValueFactory<Post, Integer>("commentCount"));
+             tclike.setCellValueFactory(new Callback<CellDataFeatures<Post, Integer>, ObservableValue<Integer>>() {
+                @Override
+                public ObservableValue<Integer> call(CellDataFeatures<Post, Integer> param) {
+                    int postId = param.getValue().getId();
+                    int likeCount = service.getLikeCount(postId);
+                    return new SimpleIntegerProperty(likeCount).asObject();
+                }
+            });
+
+               tcdislike.setCellValueFactory(new Callback<CellDataFeatures<Post, Integer>, ObservableValue<Integer>>() {
+                 @Override
+                 public ObservableValue<Integer> call(CellDataFeatures<Post, Integer> param) {
+                     int postId = param.getValue().getId();
+                     int dislikeCount = service.getDislikeCount(postId);
+                     return new SimpleIntegerProperty(dislikeCount).asObject();
+                 }
+             });   
+
+               
+               // add an event listener to the search field
+    searchField.textProperty().addListener((observable, oldValue, newValue) -> {
+        searchPosts();
+    });
+    
+                 tftableview.setItems(list);
     }
 
     @FXML
@@ -169,14 +235,33 @@ public class PostBackController implements Initializable {
         return;
     }
         else{
-        ServicePost service=new ServicePost();
-        service.deletePost(p.getId());
-         refreshtable();
-         Alert alert = new Alert(Alert.AlertType.INFORMATION);
-        alert.setContentText("Post deleted successfully");
-        alert.showAndWait();
-         }
+          ServicePost service = new ServicePost();
+          Dialog<ButtonType> confirmationDialog = new Dialog<>();
+        confirmationDialog.setTitle("Confirmation");
+        confirmationDialog.setHeaderText("Are you sure to delete this post?");
+
+        // set the graphic for the confirmation dialog window
+        Stage stage = (Stage) confirmationDialog.getDialogPane().getScene().getWindow();
+        stage.getIcons().add(new Image("/image/logo.png"));
+
+        ButtonType confirmButton = new ButtonType("Confirm", ButtonBar.ButtonData.OK_DONE);
+        ButtonType cancelButton = new ButtonType("Cancel", ButtonBar.ButtonData.CANCEL_CLOSE);
+        confirmationDialog.getDialogPane().getButtonTypes().setAll(confirmButton, cancelButton);
+
+        Optional<ButtonType> result = confirmationDialog.showAndWait();
+        if (result.isPresent() && result.get() == confirmButton) {
+            // delete the post
+            service.deletePost(p.getId());
+            refreshtable();
+
+            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+            alert.setContentText("Post deleted successfully");
+            alert.showAndWait();
+        }
+
     }
+    }
+        
     @FXML
    public void redirectToUpdate(javafx.event.ActionEvent event) throws IOException {
        
@@ -218,6 +303,25 @@ public class PostBackController implements Initializable {
             stage.show();
         
     }
+
+    @FXML
+    private void displayStats(javafx.event.ActionEvent event) throws IOException {
+         FXMLLoader loader = new FXMLLoader(getClass().getResource("Stats.fxml"));
+            Parent root = loader.load();
+            
+            Scene scene = new Scene(root);
+            Stage stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
+            stage.setTitle("Stats Post");
+            stage.setScene(scene);
+            stage.show();
+    }
    
+    
+    public void searchPosts() {
+         ServicePost service=new ServicePost();
+    String query = searchField.getText();
+    List<Post> searchResults = service.searchByTitle(query);
+    tftableview.setItems(FXCollections.observableArrayList(searchResults));
+}
 }
